requirements.txt
# Core data science packages
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Statistical analysis
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Spatial analysis and GIS
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
folium>=0.14.0
contextily>=1.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Web scraping and API access
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0

# Data processing
openpyxl>=3.1.0
xlrd>=2.0.0
pyarrow>=12.0.0

# Jupyter and notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Additional utilities
tqdm>=4.65.0
click>=8.1.0
rich>=13.4.0

# Machine learning and advanced analytics
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Time series analysis
prophet>=1.1.0
arch>=6.2.0

# Spatial statistics
pysal>=2.8.0
esda>=0.3.0
splot>=1.1.7 